requirements.txt
streamlit
groq
requests
Pillow
PyPDF2
